requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
flask-bcrypt==1.0.1
python-dotenv==1.1.0
werkzeug==3.1.3